Remove reflective call classes and handle everything in the compiler.
The compiler already had to do some special-casing for primitives
and hijacked classes in the codegen for reflective calls. For
primitives, it relied on the presence of dedicated classes in the
run-time library.

Now, the compiler directly generates code for what used to be the
body of the target method in reflective call class. This allows to
entirely remove the reflective call classes from the library.