[no-master] Fix #3082: Refuse to generate a launcher for new-style main.
When using a new-style main class (with a `main(Array[String])`
method), it is not possible to use launchers. In particular, the
setting `scalaJSUseMainModuleInitializer` must be true.

Previously the sbt plugin was not checking this and would simply
generate a broken launcher, resulting in a run-time error. Now we
report a meaningful error when trying to generate the launcher.