Make testSuiteJVM/test pass on all system configurations.
This commit forces testSuiteJVM to use

* UTF-8 as default encoding
* en-US as default locale
* `\n` as line separator

This ensures that it succeeds on all system configurations, e.g.,
on Windows or on a system configured with a non-English default
locale.

It is necessary because Scala.js ensures those defaults "by spec",
and hence our test suite relies on these defaults, notably to test
them.