Fix #3090: Better default settings for custom sbt configurations.
Previously, some settings involving `in Compile` or `in Test` were
defined in the project-level settings. We now define them from
`scalaJSCompileSettings` and `scalaJSTestSettings`, so that they
are automatically and accurately applied when those sets of
settings are reused on custom sbt configurations (e.g.,
`IntegrationTest`).

We also make sure that those settings do not hard-code the `-test`
suffix, but rather derive it from the current configuration name.