Remove `with java.io.Serializable` from `RuntimeLong`.
`RuntimeLong` already extends `java.lang.Number`, which itself
implements `java.io.Serializable` anyway, so this is fine.

We need to do this now because 2.12.3 will anyway remove the
mention to `java.io.Serializable` from `RuntimeLong`'s class file,
so we end up with a binary incompatibility. It is cleaner to keep
source and binary in sync by preemptively remove it ourselves.

Note that `java.lang.Long` was already declared without the
explicit `with java.io.Serializable`, so this brings `RuntimeLong`
more in sync with `java.lang.Long`, which is not a bad thing.