Do It Ourselves publishing to Bintray; remove bintray-sbt.
Since we're not quite happy with the current state of bintray-sbt
(see https://github.com/sbt/sbt-bintray/issues/130), this commit
gets rid of it. Instead, we configure publishing to Bintray by
hand. This turns out to be only slightly more complex than
configuring Sonatype.

The new behavior will not automatically *release* the newly
published files. Instead, a manual intervention in the Bintray Web
UI will be necessary. This actually brings our Bintray publishing
more in line with Sonatype, so that's also a win.