Filter `scala.Dynamic` out of the interfaces of JS types.
JS types cannot implement Scala interfaces, so this is not valid.
For other Scala interfaces, the front-end has checks to prevent
this from happening. However, for `scala.Dynamic`, there is an
explicit loophole, necessary to be able to define dynamically typed
JS types.

The loophole is definitely unsound, though. Since we do not want
our IR to be unsound, we have to filter `scala.Dynamic` out from
the list of interfaces implemented by JS types in the codegen.