Fix #3128: Avoid a binary incompatibility on JDK 7.
The method `java.util.concurrent.ConcurrentHashMap.keySet()` got a
more precise result type in JDK 8. Since we build the distribution
on JDK 8, the compiler links to the more precise signature, which
then fails to link at run-time on a JDK 7. We work around the
problem by an explicit upcast to `java.util.Map`, to force the
compiler to resolve to the binary compatible bridge.