Rewrite and complete j.l.String{Builder,Buffer}.
This is a complete rewrite of `java.lang.StringBuilder`, with the
complete set of methods from JDK 8. This includes `delete()`, which
fixes #3124. The test suite has also been rewritten.

`java.lang.StringBuffer` is also rewritten, but delegates all its
operations to an underlying `StringBuilder`. Since this class is
rarely used, due to its bad reputation (synchronization), it is not
worth making it fast.

A few methods of `java.lang.String` are also added, if they were
necessary to implement the corresponding ones in `StringBuilder`.
No effort was made to complete the set of methods of `String` in
this commit.