[no-master] Further reduce the pressure on PhantomJS in the test suite.
Once again, we are forced to reduce the test suite on PhantomJS, as
we initially did in 8166c6c782a13ea117c01b9d0761418671b69c9b.

This time, we remove the tests for charset encoders and decoders,
as well as all tests that require specific versions of Scala or of
the JDK. Those should also not be sensitive to their environment.