Move the Int to Long conversion to `RuntimeLong.fromInt(x)`.
Previously, it was an overloaded constructor `new RuntimeLong(x)`.
This made sense in order to avoid loading the module instance of
`RuntimeLong`. However, now that `RuntimeLong` has an elidable
module accessor, this optimization is not necessary anymore.

With this change, `RuntimeLong` always has exactly one constructor,
and it can therefore benefit from the "inlined constructor"
optimization.