Avoid calling the super JS constructors in Scala classes.
With the exception of "inlined inits", the job of the JS
constructor in a Scala class is to create the fields, initialized
to the zero of their types.

Previously, the JS constructor of some class `A` extending `B`
created its fields as follows:

1. call the super constructor to create all inherited fields,
2. create the fields directly declared in `A`.

    function $c_A() {
      $c_B.call(this);
      this.aField1 = 0;
      this.aField2 = 0;
      <optional inlined init body>
    }

This was however non-optimal, as calling the super constructor is
pretty expensive. Instead, we now make the JS constructor of `A`
directly create all its fields, including inherited ones:

    function $c_A() {
      this.bField1 = 0;
      this.bField2 = 0;
      this.aField1 = 0;
      this.aField2 = 0;
      <optional inlined init body>
    }

This optimization is only performed in ECMAScript 5.1, since with
ES 2015 classes, calling the super constructor is mandatory.

This optimization is ridiculously effective, as shown by our
benchmark suite:

* Tracer is 25% faster
* DeltaBlue is 37% faster
* SHA-512 is 14% faster