Remove `scalaJSJavaSystemProperties`.
Instead, we directly derive them inside the addition to
`jsExecutionFiles`. Getting rid of that setting makes the public
API of the sbt plugin a bit nicer.

The only use case seemed to be in our own build, which patches the
Java system properties for `testHtml`. So that part needs to
duplicate a tiny bit more code from the sbt plugin to compensate.