Adapt the codebase to changes in nio Buffers in Java 9.
First, several methods of `nio.Buffer` are now overridden in
specific subclasses so that they return a more specific type. For
example, `Buffer.position(int)` is overridden in `ByteBuffer` to
return a `ByteBuffer`, allowing chained invocations.

This obviously means we had to add those overrides in our javalib,
so that they can link. Indeed, the overrides have a different
binary signature.

In addition, because of https://github.com/scala/bug/issues/10418,
we also have to call `position()` and `limit()` instead of
`position` and `limit`, respectively.

Lastly, the methods `slice()` and `duplicate()` now have an
abstract definition in `Buffer`, which broken source compatibility
of the `BufferAdapter`s in our test suite. We fix this by renaming
the methods to `sliceChain()` and `duplicateChain()` in the
adapters.