Redefine `JSSuperBracketSelect` and `JSSuperBracketCall`.
Previously, they took as argument a `ClassType` of the enclosing
class. It was up to the emitter to look up the super class, and
load its class value.

Now, they take the class value of the super class. For example, if
`Foo` extends `Bar`, they now take
`LoadJSConstructor(ClassType(Bar))` instead of `ClassType(Foo)`.

This notably allows to unify the way super selects and super calls
are handled in anonymous JS classes. The new encoding will also be
necessary for proper support of nested JS classes.