Make the lexical Env of FunctionEmitter more robust.
The lexical `Env` of `FunctionEmitter` was previously not always
complete. It was not a big deal, because it is only used to avoid
spilling expressions into temporaris if they are known to be
immutable.

However, handling class captures in nested JS classes will require
the `Env` to be complete, so that we can distinguish local
variables from class captures. It is also internally nicer to have
complete `Env`s.