Precompute information about the params of an `Exported`.
Previously, in `GenJSExports`, the effective type of parameters,
taking entering-uncurry and entering-posterasure into account, was
recomputed several times within `genExportSameArgc`, through
`computeExportArgType`.

In this commit, `ExportedSymbol` itself is responsible for
computing all the required information about its parameters, once.
This can slightly speedup the generation of overloaded exports,
but is mostly useful as a cleanup, as it gathers the relevant logic
in a single place.