Fix #2805 and #2398: Nested JavaScript classes.
This commit dramatically changes the way nested JS classes are
compiled, to fix a pervasive issue about their semantics. We
call a class "nested" if it either *inner* (defined in the body
of a class, trait, or non-static object) or *local* (defined
inside a `def`).

The semantic issue is that we have always allowed the definition
of nested non-native JS classes, but using the normal compilation
scheme of scalac, driven by explicitouter and lambdalift. This
caused their constructor to be visible to JavaScript (and JS
interop within Scala.js) as having more arguments, for the
captures. For Scala classes, this is not an issue, as their
constructor cannot be manipulated by JavaScript. But for JavaScript
classes, it is probablematic.

For example, consider:

  class Enclosing(val x: Int) {
    class InnerJSClass extends js.Object {
      def getX(): Int = x
    }
  }

Because of `explicitouter`, the constructor of `InnerJSClass` takes
an additional parameter `$outer` of type `Enclosing`. This means
that if we ask for `js.constructor[enclosing.InnerJSClass]`, we
receive a class value that is basically unusable.

For local classes, lambdalift would add zero to many parameters for
captures of local vals and vars.

For nested *native* JS classes, there was "no issue" in the sense
that they were disallowed as compile errors to begin with. However,
this is a limitation, as evidenced by the issue #2398.

The new semantics for nested JS classes are the "obvious" ones. If
they capture stuff (the outer pointer and/or local vals), they are
properly *captured* by the JS class value, and are therefore not
visible as constructor parameters.

Additionally, and relatedly, this means that for inner JS classes,
there is a different JS class value for each enclosing instance.
For local JS classes, there is a different JS class value for each
time the code "executes" the class declaration.

Together, these changes to the semantics address #2805.

Given all the infrastructure that is set up for the above, it also
becomes trivial (and more consistent) to address #2398 in the same
sweep. Indeed, it is virtually only a matter of lifting compile-time
checks.

Implementing the above semantics is far from being trivial, because
erasure tends to destroy a lot of information that we need. In
particular, it will destroy the *paths* of inner JS classes (or
*prefixes* in scalac terms). flatten will also destroy the identity
of nested JS classes, preventing from distinguishing inner classes
by their enclosing instance and local classes by their "execution
path".

We address this with actual code transformations of the scalac trees,
something that has no real precedent in our compiler plugin. We
actually reify nested JS classes, in two steps, before erasure kicks
in and destroys what we need.

The first step, in phase `xplicitinnerjs`, creates for every inner JS
class, a field in its enclosing class to hold its JS class value.
The second step, in phase `xplicitlocaljs`, does something similar
with local vals for local JS classes, and transforms all references
to preserve identities. Ideally, this would be done in a single
phase, but we need to do some work *before* the phase `fields` of
Scala 2.12+, and some *after* that phase. In particular, we need to

* create the inner JS class fields before `fields`, so that the
  latter phase will correctly mix them in subclasses of traits, and
* desugar references to module classes after `fields`, since it is
  that phase that reifies modules.

Since there can be dependencies of all sorts between *references* to
inner- and local JS classes, all references must be processed in the
same phase. This is why `xplicitlocaljs` actually desugars all
references, including those to inner JS classes. `xplicitinnerjs`
only creates the fields that will hold the JS class values.

For further information on the implementation, and code snippets
before and after the transformations, refer to the long Scaladoc
comments of those two phases.