Remove the `scripts/build-all-js.sh` script.
The large number of Scala versions to support, combined with the
size of the test suite and the memory-hungry linker and optimizer
means that this script cannot reasonably be used anymore. It is
easier to do things by hand.

Therefore, this script is just another thing we need to update each
time we upgrade to a new Scala version, for no benefit.