Move the primitive `js.Object.hasProperty` to `js.special.in`.
The main reason is that `js.Object` is a JS type, and JS type
should not contain primitives, as the semantics of such primitives
is at odds with the more general semantics of JS types.

It is also more consistent with `instanceof`, the other alphabetic
binary operator of JavaScript.

As a nice benefit, we sidestep the issue with the evaluation order
of the arguments, and get support for symbols.

The method `js.Object.hasProperty` is preserved as a user-defined
implicit pimp on `js.Object.type`.