Fix #1265: Intrinsify `js.Dictionary.apply` and `js.Dynamic.literal`.
To be more exact, they are both written in user-space as forwarders
to a new method `js.special.objectLiteral`, which is itself an
intrinsic.

This improves the generated code for calls to `js.Dictionary.apply`
with known variadic arguments. It will also improve the generated
code for user-defined helpers that forward themselves to
`js.Dictionary.apply` or `js.Dynamic.literal`.