Actually use the implementations of hijacked classes in the emitter.
Previously, the instance methods of hijacked classes were thrown
away by the emitter. Instead, we relied on a duplicate
implementation of those methods in the form of helpers in
`scalajsenv.js`. Those helpers were doing both the dispatch that
would normally be done by method calls, and the actual body of the
methods.

Now, the body of the methods of hijacked classes is generated using
normal method emission. They take their `this` argument as an
explicit parameter, just like default methods. The helpers in
`scalajsenv.js` only do the dispatch, and delegate to those
generated "hijacked methods".

In addition, for monomorphic calls (when the receiver of an
`Apply` is known to be a specific hijacked class), the function
emitter generates code that directly calls the hijacked method,
bypassing run-time dispatch. This means that the methods that
can *only* be called monomorphically, such as `$isNaN`, do not
need to be in `scalajsenv.js` at all anymore.

The fact that the bodies are not written in `scalajsenv.js` means
that the Emitter does not need to explicitly declare symbol
requirements for those bodies. This has only marginal effect, but
it does allow to get rid of `Double$.compare__D__D__I`.

The fastopt.js size does increase because the methods for
`toString`, `hashCode` and `equals` of all hijacked classes are
emitted, although almost none of them are actually ever called.
GCC can get rid of those, though, so it does not impact opt.js.