Bye bye RuntimeString: use a complete String class in the IR.
Previously, the `String` class in the IR was truncated to a very
small subset of its methods. It only had the methods that override
or implement methods from its ancestors. Those were critical
because the class hierarchy would otherwise be bogus.

All other methods, however, were only part of `RuntimeString`
(including static methods). It was the compiler's responsibility
to rewire calls to those methods to `RuntimeString`.

We had done this because all the methods in hijacked classes had
to have a duplicated implementation in hand-written JavaScript, in
`scalajsenv.js`. It would have been not only tedious to duplicate
the large number of methods of `String` in `scalajsenv.js`, but
also would have had serious consequences on DCE, given that all
the methods in `scalajsenv.js` have to be considered reachable.

Now, the parent commit has changed the rules of the game: only
methods that are potentially polymorphic need a dispatcher in
`scalajsenv.js`. All other methods are generated from the IR.
This means that we can implement all these other methods of
`String` only once in the IR, and not worry about JavaScript code
or DCE.

In this commit, we therefore make `String` more like the other
hijacked classes, having the implementations of all its methods,
written in a proper .scala source file in `javalanglib/` (as
opposed to having hard-coded IR in the build).

There is a small catch, though: scalac does not like to be given
`String` to compile. Therefore, we write the class and its
companion object as if they were named `_String`. We add a piece of
magic in the compiler to rename `java.lang._String` into
`java.lang.String` when generating its IR. One would think this
makes the new situation no better than what we had before
(special-casing `String` in the compiler) but there is a
fundamental difference: we only need this magic when compiling
`String.scala` itself, not when compiling clients of `String`. The
magic is therefore contained within our build, instead of being
specified in the public definition of the IR.