Uniformize handling of constructors of hijacked classes.
Now the constructors of all hijacked classes are handled in the
way `String`'s were, i.e., they each have a corresponding static
method in the companion module. That method is called `new` in all
hijacked classes, which is a pretty safe name to use as the JDK is
unlikely to ever add a method named `new`.