Move the fround polyfill to an implementation in JavaScript.
This move does obfuscate somewhat the implementation, since we have
to incorporate the optimized version rather than relying on our
optimizer. However, it has three benefits.

First, it was awkward that `froundPolyfill` had to return a
`Double` whereas it was really returning a `Float`. This was a
direct consequence of the fact that a high-level Scala method was
used to implement a very low-level primitive operation.

Second, the other polyfills for ECMAScript 2015 that our emitter
relies on (`imul` and `clz32`) were already implemented in
JavaScript. It makes more sense to be consistent in that aspect.

Finally, it removes the last accidental dependency of the emitter
on a non-JDK class. This means that the emitter is less dependent
on Scala as a source language.