Move `runtime.Bits` to `java.lang.FloatingPointBits`.
The main role of `Bits` is to manipulate the bits of
floating-point numbers, and the only consumers of that API were
`java.lang.Float` and `java.lang.Double`.

Two places (nio Buffers and `testsuite.utils.Platform`) reused some
of the detection of typed arrays performed in `Bits`, but it makes
more sense that they do it on their own.