Do not inline `.toJSArray` to prevent code size blowing up.
For an explicit `.toJSArray`, it is typically not advantageous to
inline the method unless the exact type of collection is known.
And in most of those cases, the collection would be
stack-allocated, which would force inline anyway.

In this commit, we directly implement `.toJSArray` (instead of
depending on a method of `runtime`) and mark it as non-inlined.

The method `runtime.genTraversableOnce2jsArray`, used by the code
gen for JS varargs, is unchanged.