Consolidate `NullType` and `NothingType` as primitive types.
They receive the one-letter names `N` for "Null" and `E` for
"Empty type" or "throws Exception", respectively. Those letters
designate the `ClassRef`s for `NullType` and `NothingType`, and
are used in encoded names of methods that take such types as
parameters or as result types. They are accessible as
`ir.Definitions.NullClass` and `ir.Definitions.NothingClass`.

Similarly to `VoidClass`, arrays of `N` or `E` are illegal, and the
IR checker now checks that they are not used. Moreover, and unlke
`VoidClass`, even `classOf[N]` and `classOf[E]` are illegal.

This means that `Array[Null]` and `Array[Nothing]` in Scala source
code keep being encoded as `Array[scala.runtime.Null$]` and
`Array[s.r.Nothing$]`, and similarly for `classOf[Null]` and
`classOf[Nothing]`. This retains compatibility with Scala/JVM.

After the changes in this commit, the IR and the Linker do not know
about `s.r.Null$` and `s.r.Nothing$` anymore.