[no-master] Turn the compiler crash of #3211 into a linking error.
At the very least, the compiler should not crash on any valid
Scala code, so this commit makes the situation a bit less bad than
it was before.

It does not really *solve* the issue, though, as there is still a
linking error. That, however, seems difficult if not impossible to
properly fix, as explained in #3211, so we do not attempt to
address it now.