Reintroduce tests for our EMCAScript 2015 polyfills.
When we removed support for Rhino, and moved the support of
PhantomJS in a separate repository, we accidently got rid of all
the tests in our build checking our polyfills for ES 2015 features.
For example, the polyfills for `Math.imul`, or the floating-point
bits conversions (including `Double.hashCode`).

This commit restores that by introducing in our build a custom
`NodeJSEnvForcePolyfills`, which is like `NodeJSEnv` but forcefully
`delete`s all the good stuff from ES 2015.

This ensures that we do not introduce regressions in our polyfills
in the future.