Restore the specified behavior of j.l.Class.isAssignableFrom.
The implemented behavior used to match the specification, but was
changed in 9108b7c96c5b6fd7881c69a738dcc9ab4a7ef1df to fix #879.

It turns out that the real culprit for #879 was not
`isAssignableFrom`, but the implementation of `ClassTag.unapply`.
Now that the latter has a Scala.js-friendly implementation in all
versions of Scala for which it is not also broken on the JVM, we
do not need `isAssignableFrom` to work around it.

This makes sure that the behavior of `java.lang.Class` is not
dependent on Scala as a source language.

One partest does not pass out-of-the-box in 2.11.x if we do that:
`run/t6318_primitives.scala`, but it works in 2.12.x. That happens
because 2.12.x removed the overloads taking primitives, so at a
source call site of `ct.unapply(5)`, 2.11.x would call
`unapply(Int)` but 2.12+ would call `unapply(Any)`. It turns out
that the `unapply`s taking primitives are not Scala.js-friendly
enough for this partest (which has an overridden check file, btw).
We fix this by "backporting" from 2.12 the removal of the
`unapply`s taking any. In practice we cannot do that because we
need to preserve binary compatibility, but we can implement it
anyway by overriding them to delegate to the `unapply` taking
`Any`.