Implement `isAssignableFrom` in JS instead of exposing `getFakeInstance`.
Creating a fake instance and using it as an argument of
`isInstance` is a huge hack. It's OK to do it internally in
`scalajsenv.js`, as long as fake instances do not leak. But
exposing fake instances to the Scala.js code is extremely
dangerous, as it means leaking implementation-specific details at
the IR level, which could be (mis-)processed by the optimizer.

Since `getFakeInstance` was used only for `isAssignableFrom`
anyway, it is much cleaner to directly implement `isAssignableFrom`
in `scalajsenv.js`, and not expose `getFakeInstance` at all.