Change the signature of `Class`'s constructor to take an `Object`.
This way, the "public" signature of that constructor does not
refer to `ScalaJSClassData`, which is an implementation detail of
`java.lang.Class`. It is public because it is accessed by the
`Emitter`, even though it is declared as private.

Note that even though `ScalaJSClassData` as such is an
implementation detail, its *structure* must be considered as
specified at the interface between the `java.lang` implementation
and the linker. If someone wants to write a different linker, or a
different `java.lang` library, they must adhere to that structure
to ensure interoperability.