Drop support for Scala 2.10.x.
In this commit, we completely drop support for Scala 2.10.x in
Scala.js code. Note that our JVM artifacts are still cross-compiled
with 2.10.x, because we still support sbt 0.13.16+ which runs on
2.10. However, it will not be possible to use
`enablePlugins(ScalaJSPlugin)` with `scalaVersion := "2.10.x"`
anymore.

This includes:

* Drop CI tasks using 2.10.x with JS artifacts (but not with JVM
  artifacts).
* Remove all workarounds for Scala 2.10 compiler bugs, such as
  crashes for some usages of `js.Dynamic`.
* Remove all 2.10-specific stuff from `CompatComponent` in the
  compiler.
* Remove all 2.10-specific assumptions or branches in the test
  suite.

This commit does not include:

* Drop `js.Dynamic.x` and the associated compiler tricks, which
  work around the 2.10 *library* (not compiler), and deserve a
  separate commit anyway.
* Take advantage of new 2.11 language features, such as making the
  underlying val of value classes private.

Those items are left for further commits.