Fix #3219: Correctly recognize exports for `$xyz` in the back-end.
We were erroneously using `sym.unexpandedName` when testing whether
a method is an export bridge. However, if the export name happens
to start with a `$`, then `unexpandedName` considers the whole
thing as an expanded name, and removes the whole prefix as if it
were the result of an expansion.

Simply using the `name` for the test fixes the issue. Note that
`jsExportInfo` was already using `name`, and not `unexpandedName`.