Fix #3218: Do not execute `main()` in `Test` twice for `test:run`.
Running `test:run` depends on `loadedJSEnv in Test`, which itself
triggered the discovery of tests, resulting in an additional
instance of the .js file to be executed. This meant that the
`main()` was effectively run twice when doing `test:run`.

We fix this problem by backporting one change from Scala.js 1.x,
which is to trigger test discovery only on `loadedTestFrameworks`,
but not on `loadedJSEnv`.