Fix #3228: Disable reflective instantiation for nested modules.
We were using `sym.isModuleClass` to identify module classes when
generating the reflective instantiation code, but that also matches
nested objects. Now, nested objects are not compiled as module
classes (only static objects are) because there is one instance per
enclosing instance. This caused linking errors as soon as the
codebase contained a nested object with
`@EnableReflectiveInstantiation`.

Since non-static objects are not really *loadable* as per the API
of `Reflect` (we would need a reference to the enclosing instance),
we fix this issue by not registering non-static objects at all.