Make UndefinedBehaviorError extend VirtualMachineError.
`VirtualMachineError`s is the superclass of severe exceptions that
indicate the virtual machine has reached a state that is
potentially invalid, and cannot be safely recovered from. This is a
perfect description for `UndefinedBehaviorError`s.

Since `VirtualMachineError`s are considered fatal by Scala (i.e.,
not `NonFatal`), this allows `UndefinedBehaviorError` not to extend
`scala.util.ControlThrowable`, which was a hack to begin with, and
therefore not depend on the Scala standard library.

In the process, we harmonize the `message` of
`UndefinedBehaviorError` with other `VirtualMachineError`s, as
mentioning both the class name and the message of the underlying
exception.