User-space implementation of `js.{Any,ThisFunction}.fromFunctionN`.
Previously, those methods were primitives. They did have a user-
space body, but it was not used. This was done because most JS
functions were the result of applying those implicit conversions.
However, now that we have SAM support in all the versions of Scala
for our library, we can get away without them being primitives.

If the user's codebase has SAM support, this commit changes
virtually nothing. If it doesn't (it uses Scala 2.11 without
`-Xexperimental`), then this commit produces slightly worse code
than before (there is a useless double-box of functions), but it is
still correct.

If we assume that most end-user codebases should use Scala 2.12
anyway, we are not losing much. We could also improve the optimizer
to remove the double-box in the future, if necessary. This commit
does not do that, however.