Remove the `js.Dynamic.x` workaround.
The explicit presence of `x` in `js.Dynamic` was a workaround for
the implicits in `Predef` that were accidentally providing `x` on
`Any`.

Since those implicits were removed in 2.11, and we do not support
2.10 anymore, we can remove the workaround.

This surfaces a spectacular scalac bug in one of our tests, which I
have not been able to minimize. The problem was already visible
with `y` or any other field not named `x`, so the explicit
`js.Dynamic.x` was hiding another bug.