Use a "minilib" for the unit tests of the linker.
Instead of using the full `scalajs-library.jar` for the unit tests
of the linker, we now use a minimal library. That "minilib"
contains things like `Object`, `Integer` and `RuntimeLong`: core
classes without which it is not possible to link a hello world.

The minilib does not contain `js.Any` nor any of its subtypes,
proving that the linker can live without them. That required to
tweak a few core methods not to use niceties from our standard
library, such as `JSStringOps` and `JSNumberOps`. Instead, we must
be careful to only use `js.Dynamic`, and only inside methods, so
that even `js.Dynamic` does not leave any trace in method
signatures.