Fix #3210: First-class private methods in the IR.
Previously, private methods were modeled in the IR as normal
methods whose name had a special encoding to ensure that they were
non-ambiguous across their entire hierarchy.

In this commit, we mark private methods using an explicit flag,
similar to static methods. Those combine as the *namespace* of a
member, which can be private or not, static or not, and
constructor or not. Each namespaces can define methods of the same
name (with signature), without them being ambiguous.

Only the public non-static namespace is *inherited* by subclasses.
Private and/or static methods, as well as constructors, can only
be accessed using their immediately enclosing class. Therefore,
private methods and constructors can only be called using
`ApplyStatically`, which encodes the enclosing class.

In the emitter, private methods and constructors are encoded
similarly to default methods: they are top-level functions taking
an explicit `this` parameter.

In this commit, constructors must always be public, which
corresponds to how things were done previously (private
constructors had an encoded name without the `__pXyz` marker). We
could possibly revisit this in a future commit, by allowing
private constructors in the IR as well.

Static initializers are encoded in the public static constructor
namespace.