Separate `tools` into `io`, `logging` and `linker`.
They define the packages `tools.io`, `tools.logging` and
`tools.linker`, respectively.

This allows `js-envs` and related artifacts, as well as
`test-adapter`, not to depend on the `linker` nor `ir` anymore.
Therefore, those artifacts can be used without bringing the linker
and IR into the classpath, which can be very important for build
tools that want to load the linkers of various Scala.js versions in
separate class loaders.