Introduce a true for-in loop in the IR.
This commit adds first-class support for the general `for-in` loop.
It can be produced with `js.special.forin`, which is now used to
implement `js.Object.properties`.

This is important because `js.Object.properties` cannot encode all
uses of `for-in` loops. While it can encode most of the semantics
applying to ordinary JS objects, it is not enough for exotic
objects, notably proxies. Only exposing `js.Object.properties` was
therefore a hole in our interoperability features.

In addition, if another language wants to compile to the IR, they
should have a way to encode an efficient `for-in` loop, which was
previously not possible, as the only efficient `for-in` loop was
based on intrinsifying a method from the Scala.js standard library.