Make `js.FunctionN`s map to arrow functions in ECMAScript 2015.
Arrow functions have some distinct semantics besides not capturing
`this`: they cannot be *constructed* (i.e., called with `new`) and
they do not have a `prototype` property. Therefore, they also
cannot be used as "parent classes".

As such, we need an interop feature to create true arrow functions
when targeting ES 2015+. The existing `js.FunctionN`s are actually
a perfect fit for this, since they already do not have access to
their `this` parameter.

This is, in theory, a breaking change, as someone could have relied
on calling a `js.FunctionN` with `new` (which succeeded before, and
now throws). However, without access to the `this` value, there is
no non-contrived use case for that, so it is unlikely to be
noticeable.