Consolidate the notion of transient node in the IR.
Previously, we had two kinds of nodes in the IR used for internal
purposes: `UndefinedParam` and `CallHelper`. Those nodes should not
be part of the public IR definition.

We get rid of that public display by introducing a unique, generic
`Transient` node, which can be used by internal phases for custom
purposes. The scalac plugin uses it for `UndefinedParam`, and the
optimizer+emitter use it for `CallHelper`. Transient nodes could be
used by any other phase, including user-defined tools manipulating
IR trees.