Stricter typing for `JSSpread`s in the IR.
Previously, `JSSpread` extended `Tree` That was nonsense because
`Tree`s are supposed to be valid anywhere in statement/expression
position (provided they typecheck), although `JSSpread`s are only
valid in very specific positions, structurally speaking.

We reify those restrictions in the types of the IR nodes, by
introducing `TreeOrJSSpread` as a common super type for `Tree` and
`JSSpread`.