Do not crash the compiler after "Conflicting properties and methods".
When there are conflicting properties and methods in a non-native
JS class, `GenJSExports` returns a `js.Skip()` after reporting the
error. This `js.Skip()` previously caused `genAnonSJSDefinedNew` to
crash with an "Unexpected tree".

We now ignore `js.Skip`s at that point so that compile errors are
not followed by crashes.