Fix #3259: Generate overriding bridges in synthesized SAM wrappers.
Just like Scala/JVM needs to instruct the `LambdaMetaFactory` to
generate bridges for methods overridden by the SAM, we need to
directly generate appropriate bridges in our synthesized SAM
wrappers.

Instead of generating actual *bridges*, in the sense that they call
the canonical implementation of the method, we generate bodies for
all the bridges just like for the canonical method. This is fine
because the bodies are just calling the closure stored in a field
of the instance anyway.